feat: Set up BFF with basic endpoints
We want to handle the calls to the service with a Backend For Frontend
using Next.js routing capability.
To do so, we need to set up the api in the app folder.

This commit creates one health endpoint to check the status of the
server and another one that would communicate with the houses service
to return the houses entities.