feat: Set up initial house fetching logic
We want to create an infinite scrolling page. To accomplish that first
we need to create the structure we will use to set up the queries.

This commit sets up the query client we will use to trigger the queries.
It also creates the methods to call the proper API.
Finally, it updates the different components to show a basic list with
the fetched data.