feat: Add load more button functionality
We want to have an infinite scroll home page that fetches the next
page automatically.
To accomplish this first we need to be able to fetch the pages
manually.

This commit adds the methods and interactions to be able to fetch
pages manually by clicking a load more button at the bottom of the
list.