chore: Extract BFF URL to an env file
We want to use env files to configure different environments.
We won't be storing any secrets so we are good to commit it for now.
If we need to store secrets we should do that with other tools like
GCP secrets or ensure we don't commit them to the repo.

This commit extracts the BFF url to an env file. It also changes the
git ignore settings to let us commit the file.