fix: Next page availability miss calculation
We were not checking properly if there was a next page to fetch.
This was causing the query to refetch even if there were no more pages.
We also needed to increment the amount of elements fetched by default
to ensure we have a smoother experience when scrolling.

This commit fixes the next page issue by implementing the proper logic
to check if the current page is the last one or not.
It also adds the needed pieces to fetch a larger amount of elements
per page.