fix: Change observer position
We had the observer to trigger fetching the next page in the last
element of the list. This causes the users to get to the bottom and
wait for the fetching to finish.

This commit moves the observer to the middle of the list instead to
ensure users have a smoother experience when scrolling since the
fetch will be triggered when they reach the middle of the page.