feat: Cache data in the client and react-query
To give a user a better experience we can cache the list with the
current state. We need to ensure that we bust the cache when it's
stale. For now, we set a 12 hours time limit for it.

We also can restore this cache when the user refreshes the page to
avoid loosing the data they already fetched.