feat: Restore scroll position
To ensure the user can go over the list of items without issues, we
want to allow them to keep the page or position where they are when
refreshing just in case they have to.

This commit adds the logic to implement storing the position in the
list and retrieving it when they reload the page. Also since we
have cached data, we can ensure the data is there.